requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Gemini AI
google-generativeai>=0.3.0

# LangGraph (for agentic workflow)
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Database (Supabase)
supabase>=2.0.0

# Resume parsing
pdfplumber>=0.10.0

# GitHub analysis
PyGithub>=2.1.0

# Web scraping (India job scraper)
httpx>=0.26.0
beautifulsoup4>=4.12.0
requests>=2.31.0
crawl4ai>=0.7.0
playwright>=1.48.0

# Utilities
aiofiles>=23.2.0
python-multipart>=0.0.6
jinja2>=3.1.0
email-validator>=2.0.0
